# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lessonlink-backend バックエンドAPI",
        "title": "lessonlink-backend",
        "contact": {},
        "license": {
            "name": "typedef-tokyo."
        },
        "version": "1.0"
    },
    "host": "localhost:3002",
    "basePath": "/",
    "paths": {
...
        "/schedule/{schedule_id}/duplicate": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール複製",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schedule/{schedule_id}/item-divide": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール編集アイテムリスト分割",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "スケジュール保存リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ScheduleItemDivideRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleItemEditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schedule/{schedule_id}/item-join": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール編集アイテムリスト分割",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "アイテム結合リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ScheduleItemJoinRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleItemEditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/schedule/{schedule_id}/item-move": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "スケジュール編集アイテム移動",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ScheduleID",
                        "name": "schedule_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "アイテム移動リクエスト",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/controller.ScheduleItemMoveRequestData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/presenter.ScheduleItemEditResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "controller.ScheduleItemDivideRequestData": {
            "type": "object",
            "required": [
                "divide_minutes",
                "history_index",
                "identifier",
                "lesson_id"
            ],
            "properties": {
                "divide_minutes": {
                    "type": "integer"
                },
                "history_index": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                }
            }
        },
        "controller.ScheduleItemJoinRequestData": {
            "type": "object",
            "required": [
                "history_index",
                "join_from_identifier",
                "join_to_identifier"
            ],
            "properties": {
                "history_index": {
                    "type": "integer"
                },
                "join_from_identifier": {
                    "type": "string"
                },
                "join_to_identifier": {
                    "type": "string"
                }
            }
        },
        "controller.ScheduleItemMoveRequestData": {
            "type": "object",
            "required": [
                "duration",
                "end_time_hour",
                "end_time_minutes",
                "history_index",
                "identifier",
                "item_tag",
                "lesson_id",
                "room_index",
                "start_time_hour",
                "start_time_minute"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "end_time_hour": {
                    "type": "integer"
                },
                "end_time_minutes": {
                    "type": "integer"
                },
                "history_index": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "item_tag": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "room_index": {
                    "type": "integer"
                },
                "start_time_hour": {
                    "type": "integer"
                },
                "start_time_minute": {
                    "type": "integer"
                }
            }
        },
...
        "presenter.ScheduleItemEditLessonItem": {
            "type": "object",
            "required": [
                "duration",
                "identifier",
                "lesson_id",
                "lesson_name"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                }
            }
        },
        "presenter.ScheduleItemEditResponse": {
            "type": "object",
            "required": [
                "history_index",
                "lesson_item_list",
                "room_lesson_list"
            ],
            "properties": {
                "history_index": {
                    "type": "integer"
                },
                "lesson_item_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleItemEditLessonItem"
                    }
                },
                "room_lesson_list": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/presenter.ScheduleItemEditRoomLesson"
                    }
                }
            }
        },
        "presenter.ScheduleItemEditRoomLesson": {
            "type": "object",
            "required": [
                "duration",
                "end_time_hour",
                "end_time_minutes",
                "identifier",
                "item_tag",
                "lesson_id",
                "lesson_name",
                "room_index",
                "start_time_hour",
                "start_time_minutes"
            ],
            "properties": {
                "duration": {
                    "type": "integer"
                },
                "end_time_hour": {
                    "type": "integer"
                },
                "end_time_minutes": {
                    "type": "integer"
                },
                "identifier": {
                    "type": "string"
                },
                "item_tag": {
                    "type": "string"
                },
                "lesson_id": {
                    "type": "integer"
                },
                "lesson_name": {
                    "type": "string"
                },
                "room_index": {
                    "type": "integer"
                },
                "start_time_hour": {
                    "type": "integer"
                },
                "start_time_minutes": {
                    "type": "integer"
                }
            }
        },
...
    }
}